test: board specific test files
