+++ test/sketches/empty.ino
@@ -1,0 +1,3 @@
+void setup() {}
+
+void loop() {}
