+++ test/sketches/hello-world.ino
@@ -1,0 +1,6 @@
+void setup() {
+    Serial.begin(9600);
+    Serial.println("Hello World");
+}
+
+void loop() {}
