+++ test/sketches/mcu_s2/tof-distance-display.ino
@@ -1,12 +1,12 @@
-// Code generated by senseBox Blockly on Wed Nov 13 2024 10:48:22 GMT+0100 (Central European Standard Time)
+// Code generated by senseBox Blockly on Wed Nov 13 2024 14:34:47 GMT+0100 (Central European Standard Time)
 
 #include <SPI.h>
 #include <Wire.h>
 #include <Adafruit_GFX.h> // http://librarymanager/All#Adafruit_GFX_Library
 #include <Adafruit_SSD1306.h> // http://librarymanager/All#Adafruit_SSD1306
-#include <vl53l8cx_class.h>
+#include <vl53l8cx.h>
 
-VL53L8CX sensor_vl53l8cx_top(&Wire, -1, -1);
+VL53L8CX sensor_vl53l8cx(&Wire, -1, -1);
 
 #define SCREEN_WIDTH 128
 #define SCREEN_HEIGHT 64
@@ -19,10 +19,10 @@ float getVl53l8cxMin() {
   uint8_t NewDataReady = 0;
   uint8_t status;
 
-  status = sensor_vl53l8cx_top.vl53l8cx_check_data_ready(&NewDataReady);
+  status = sensor_vl53l8cx.check_data_ready(&NewDataReady);
 
   if ((!status) && (NewDataReady != 0)) {
-    sensor_vl53l8cx_top.vl53l8cx_get_ranging_data(&Results);
+    sensor_vl53l8cx.get_ranging_data(&Results);
     float min = 10000.0;
     for(int i = 0; i < VL53L8CX_RESOLUTION_8X8*VL53L8CX_NB_TARGET_PER_ZONE; i++) {
       if((&Results)->target_status[i]!=255){
@@ -45,11 +45,11 @@ void setup() {
 
   Wire.begin();
   Wire.setClock(1000000); //Sensor has max I2C freq of 1MHz
-  sensor_vl53l8cx_top.begin();
-  sensor_vl53l8cx_top.init_sensor();
-  sensor_vl53l8cx_top.vl53l8cx_set_ranging_frequency_hz(30);
-  sensor_vl53l8cx_top.vl53l8cx_set_resolution(VL53L8CX_RESOLUTION_8X8);
-  sensor_vl53l8cx_top.vl53l8cx_start_ranging();
+  sensor_vl53l8cx.begin();
+  sensor_vl53l8cx.init();
+  sensor_vl53l8cx.set_ranging_frequency_hz(30);
+  sensor_vl53l8cx.set_resolution(VL53L8CX_RESOLUTION_8X8);
+  sensor_vl53l8cx.start_ranging();
 
 }
